Moved graph util into graph directory
